% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples of using timelinevis in a Shiny app
This example is also
\href{http://daattali.com/shiny/timelinevis-demo/}{available online}.}
\usage{
runExample()
}
\description{
Run examples of using timelinevis in a Shiny app
This example is also
\href{http://daattali.com/shiny/timelinevis-demo/}{available online}.
}
\examples{
if (interactive()) {
  runExample()
}
}

