% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{setItems}
\alias{setItems}
\title{Set the items of a timeline}
\usage{
setItems(id, data)
}
\arguments{
\item{id}{Timeline id}

\item{data}{A dataframe containing the item data to use.}
}
\description{
Set the items of a timeline
}
\examples{
if (interactive()) {
library(shiny)
shinyApp(
  ui = fluidPage(
    timelinevisOutput("timeline"),
    actionButton("btn", "Change the data to yesterday")
  ),
  server = function(input, output) {
    output$timeline <- renderTimelinevis(
      timelinevis(data.frame(start = Sys.Date(), content = "Today"))
    )
    observeEvent(input$btn, {
      setItems("timeline",
               data.frame(start = Sys.Date() - 1, content = "yesterday"))
    })
  }
)
}
}

