% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinevis-package.R, R/timelinevis.R
\docType{package}
\name{timelinevis}
\alias{timelinevis}
\alias{timelinevis-package}
\title{Create a timeline visualization}
\usage{
timelinevis(data, showZoom = TRUE, listen, options, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{options}{Any extra initialization options
any function argument must be wrapped in htmlwidgets::JS()
http://visjs.org/docs/timeline/#Configuration_Options}

\item{width}{Fixed width for timeline (in css units). Ignored when used in a
Shiny app -- use the \code{width} parameter in \code{timelinevisOutput}.
It is recommended to not use this parameter since the widget knows how to
adjust its width automatically.}

\item{height}{Fixed height for timeline (in css units). It is recommended to
not use this parameter since the widget knows how to adjust its height
automatically.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Ignored when used in a Shiny app.}
}
\value{
A timeline visualization \code{htmlwidgets} object
}
\description{
\code{timelinevis} lets you create rich and fully interactive timeline visualizations.
Timelines can be included in Shiny apps and R markdown documents, or viewed
from the R console and RStudio Viewer. Includes an extensive API to
manipulate a timeline after creation, and supports getting data out of the
visualization into R. Based on the \href{http://visjs.org/}{'vis.js'} Timeline
module and the \href{http://www.htmlwidgets.org/}{'htmlwidgets'} R package.
\cr\cr
View a \href{http://daattali.com/shiny/timelinevis-demo/}{demo Shiny app}
or see the full \href{https://github.com/daattali/timelinevis}{README} on
GitHub.
}
\examples{
timelinevis()
timelinevis()

}
\seealso{
\href{http://daattali.com/shiny/timelinevis-demo/}{Demo Shiny app}
}

