% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timevisData}
\alias{timevisData}
\title{Timevis sample data}
\format{A data frame with 11 rows and 6 variables.}
\usage{
timevisData
}
\description{
A dataset containing sample time schedule data for a community center
that can be rendered by timevis.
}
\keyword{datasets}
