% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timevisDataGroups}
\alias{timevisDataGroups}
\title{Timevis sample group data}
\format{A data frame with 3 rows and 2 variables.}
\usage{
timevisDataGroups
}
\description{
A dataset containing groups data to be used with the \code{timevisData} data.
}
\keyword{datasets}
