% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zoomOut}
\alias{zoomOut}
\title{Zoom in the current visible window}
\usage{
zoomOut(id, percent = 0.5)
}
\arguments{
\item{id}{Timeline id or a \code{timevis} object (the output from \code{timevis()})}

\item{percent}{The amount to zoom in. Must be a number between 0 and 1. A value of 0.5
means that after zooming out the timeline will show 50\% more content.}
}
\description{
Zoom in the current visible window
}
\examples{

timevis() \%>\%
  zoomOut()

if (interactive()) {
library(shiny)
shinyApp(
  ui = fluidPage(
    timevisOutput("timeline"),
    sliderInput("zoom", "Zoom out by", min = 0, max = 1, value = 0.5, step = 0.1),
    actionButton("btn", "Go")
  ),
  server = function(input, output) {
    output$timeline <- renderTimevis(
      timevis()
    )
    observeEvent(input$btn, {
      zoomOut("timeline", percent = input$zoom)
    })
  }
)
}
}
