% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{addItem}
\alias{addItem}
\title{Add a single item to a timeline}
\usage{
addItem(id, data)
}
\arguments{
\item{id}{Timeline id}

\item{data}{A named list containing the item data to add.}
}
\description{
Add a single item to a timeline
}
\examples{
if (interactive()) {
library(shiny)
shinyApp(
  ui = fluidPage(
    timelinevisOutput("timeline"),
    actionButton("btn", "Add item today")
  ),
  server = function(input, output) {
    output$timeline <- renderTimelinevis(
      timelinevis()
    )
    observeEvent(input$btn, {
      addItem("timeline", list(start = Sys.Date(), content = "Today"))
    })
  }
)
}
}

