% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run examples of using timevis in a Shiny app}
\usage{
runExample()
}
\description{
This example is also
\href{http://daattali.com/shiny/timevis-demo/}{available online}.
}
\examples{
if (interactive()) {
  runExample()
}
}
