% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinevis.R
\name{timelinevis}
\alias{timelinevis}
\title{<Add Title>}
\usage{
timelinevis(data, showZoom = TRUE, listen = NULL, width = NULL,
  height = NULL, elementId = NULL, ...)
}
\description{
<Add Description>
}
\details{
... any function argument must be wrapped in htmlwidgets::JS()
}

