% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinevis-package.R, R/timelinevis.R
\docType{package}
\name{timelinevis}
\alias{timelinevis}
\alias{timelinevis-package}
\title{Create a timeline visualization}
\usage{
timelinevis(data, showZoom = TRUE, listen = NULL, width = NULL,
  height = NULL, elementId = NULL, ...)
}
\description{
\code{timelinevis} lets you create rich and fully interactive timeline visualizations.
Timelines can be included in Shiny apps and R markdown documents, or viewed
from the R console and RStudio Viewer. Includes an extensive API to
manipulate a timeline after creation, and supports getting data out of the
visualization into R. Based on the \href{http://visjs.org/}{'vis.js'} Timeline
module and the \href{http://www.htmlwidgets.org/}{'htmlwidgets'} R package.
\cr\cr
View a \href{http://daattali.com/shiny/timelinevis-demo/}{demo Shiny app}
or see the full \href{https://github.com/daattali/timelinevis}{README} on
GitHub.
}
\details{
... any function argument must be wrapped in htmlwidgets::JS()
}

